Updated altset template
This vector-based implementation was previously moved from ./set when
it was verified that the pointer based version seems viable with
current compilers.  (And the pointer based has various inherent
advantages, not the least of which is a closer correspondence to C++
STL implementation.)

Some files are the same for both set and alt set, and these changes
reflect that, though perhaps not as explicitly as could be hoped.
"set" is the main template with "altset" including files from ./set
where appropriate.